Moved database version management to module
