+++ cellphonedb/cellphonedb_cli.py
@@ -36,13 +36,12 @@ method.add_command(method_terminal_commands.statistical_analysis)
 method.add_command(method_terminal_commands.analysis)
 query.add_command(query_terminal_commands.find_interactions_by_element)
 query.add_command(query_terminal_commands.get_interaction_gene)
-query.add_command(query_terminal_commands.cpdb_data_report)
 plot.add_command(plot_terminal_commands.dot_plot)
 tools.add_command(tools_terminal_commands.generate_genes)
 tools.add_command(tools_terminal_commands.generate_interactions)
 plot.add_command(plot_terminal_commands.dot_plot)
 plot.add_command(plot_terminal_commands.heatmap_plot)
-tools.add_command(tools_terminal_commands.download_database)
+tools.add_command(tools_terminal_commands.download)
 tools.add_command(tools_terminal_commands.list_versions)
 tools.add_command(tools_terminal_commands.collect)
