+++ cellphonedb/src/exceptions/NoReleasesException.py
@@ -1,0 +1,2 @@
+class NoReleasesException(Exception):
+    pass
